Make amountPaid optional and rename DeterminationDecisionType
