+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/claimantresponse/ClaimantResponse.java
@@ -3,12 +3,10 @@ package uk.gov.hmcts.cmc.domain.models.claimantresponse;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import lombok.EqualsAndHashCode;
-import lombok.Getter;
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
 
 import java.math.BigDecimal;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
+import java.util.Optional;
 
 import static uk.gov.hmcts.cmc.domain.utils.ToStringStyle.ourStyle;
 
@@ -17,16 +15,17 @@ import static uk.gov.hmcts.cmc.domain.utils.ToStringStyle.ourStyle;
     @JsonSubTypes.Type(value = ResponseAcceptation.class, name = "acceptation"),
     @JsonSubTypes.Type(value = ResponseRejection.class, name = "rejection")
 })
-@Getter
 @EqualsAndHashCode
 public abstract class ClaimantResponse {
 
-    @NotNull
-    @Min(value = 0)
     private final BigDecimal amountPaid;
 
     public ClaimantResponse(BigDecimal amountPaid) {
         this.amountPaid = amountPaid;
+    }
+
+    public Optional<BigDecimal> getAmountPaid() {
+        return Optional.ofNullable(amountPaid);
     }
 
     @Override
