+++ domain-model/src/main/java/uk/gov/hmcts/cmc/domain/models/claimantresponse/DecisionType.java
@@ -1,6 +1,6 @@
 package uk.gov.hmcts.cmc.domain.models.claimantresponse;
 
-public enum DeterminationDecisionType {
+public enum DecisionType {
     CLAIMANT,
     DEFENDANT,
     COURT
